Clean up
- Make `get_edges` `const`
- Change move_count into u16
  - I don't think we'd need any more than 16000
- Remove commented out code (just leave 1)

+++ pincerhex/src/ai.rs
@@ -19,7 +19,7 @@ pub struct HexBot {
     #[allow(dead_code)]
     allow_invalid: bool,
     swap_state: Option<SwapRole>,
-    move_count: u32,
+    move_count: u16,
     #[allow(dead_code)]
     params: MCTSParams,
 }
@@ -130,9 +130,7 @@ impl HexBot {
             }
             SwapRole::Swap => {
                 if self.state.should_swap() {
-                    // self.state.swap_pieces()?;
                     self.swap()?;
-                    // self.colour = self.colour.opponent();
                     Ok(Move::Swap)
                 } else {
                     Ok(Move::Move(self.regular_move()))
@@ -158,7 +156,6 @@ impl HexBot {
 
     pub fn swap(&mut self) -> Result<(), BotError> {
         self.colour = self.colour.opponent();
-        // self.state.swap_pieces()?;
         Ok(())
     }
 }
